// src/model/agent/enum/Droit.ts
export enum Droit {
  SIGNER = "SIGNER",
  DELIVRER = "DELIVRER",
  ATTRIBUER = "ATTRIBUER",
  METTRE_A_JOUR_ACTE = "METTRE_A_JOUR_ACTE",
  METTRE_A_JOUR_RC_RCA_PACS = "METTRE_A_JOUR_RC_RCA_PACS",
  CREER_ACTE_TRANSCRIT = "CREER_ACTE_TRANSCRIT",
  CREER_ACTE_DRESSE = "CREER_ACTE_DRESSE",
  CREER_ACTE_ETABLI = "CREER_ACTE_ETABLI",
  CREER_PACS = "CREER_PACS",
  INFORMER_USAGER = "INFORMER_USAGER",
  CONSULTER = "CONSULTER",
  CONSULTER_ARCHIVES = "CONSULTER_ARCHIVES",
  AUCUN = "AUCUN",
  SAISIR_REQUETE = "SAISIR_REQUETE",
  DELIVRER_COMEDEC = "DELIVRER_COMEDEC"
}